Move rtss FPS counters to GPU factors. Fix rtss api usage text.
