+++ LibreHardwareMonitorLib/Hardware/Gpu/AmdGpu.cs
@@ -312,6 +312,8 @@ internal sealed class AmdGpu : GenericGpu
 
     public override void Update()
     {
+        base.Update();
+
         if (_d3dDeviceId != null && D3DDisplayDevice.GetDeviceInfoByIdentifier(_d3dDeviceId, out D3DDisplayDevice.D3DDeviceInfo deviceInfo))
         {
             _gpuDedicatedMemoryTotal.Value = 1f * deviceInfo.GpuVideoMemoryLimit / 1024 / 1024;
